requirements.txt
django>=3.1.4
djangorestframework>=3.12.2
django-cors-headers==3.13.0